requirements.txt
flask==3.0.3
markdown==3.7
python-dotenv==1.0.1

Werkzeug~=3.1.3
Flask-Migrate~=4.0.7